Start/reload with unknown upstreams fix.
When nginx starts and can't resolve weave name - it fails to start.
Fix it with some hack: script run in background, check dns and save it to files. Consul-template gets settings from this files.

CMD in exec form.
CMD in shell form start commands in sh. So we don't get signals normally and we have one no needed sh process.